refactor: Oeuvre - 작품 정보 컴포넌트 분리(메인 정보 별도 분리)
