+++ src/feature/Oeuvre/components/common/style/oeuvreInfo.scss
@@ -16,7 +16,6 @@
         }    
     }
 
-
     .cover-container {
         width: 200px;
         height: 200px;
@@ -28,49 +27,7 @@
         display: flex;
         flex-direction: column;
         >:not(:first-child) {
-            padding-top: 1rem;
-        }
-    }
-
-    .main-info-container {
-        display: flex;
-        flex-direction: column;
-
-        >:not(:first-child) {
-            margin-top: 0.5rem;
-        }
-    }
-
-    .title {
-        font-size: 1.8rem;
-    }
-
-    .artist-container {
-        font-size: 1.1rem;
-        color: rgba(var(--font-color), 0.8);
-        display: flex;
-        .button {
-            background-color: transparent;
-            padding: 0 0 0 0;
-        }
-        >:not(:first-child) {
-            margin-left: 0.5rem;
-        }
-        >:not(:last-child)::after {
-            content: ","
-        }
-    }
-    
-    .genre-container {
-        font-size: 0.75rem;
-        display: flex;
-
-        .button {
-            padding-left: 8px;
-            padding-right: 8px;
-        }
-        >:not(:first-child) {
-            margin-left: 0.5rem;
+            margin-top: 1rem;
         }
     }
